Theme Create
I start the project. First, the theme created to add on project. The
ESLint rules are adding on the project because it is written with
international standards.

But I can't decide by a file structure yet. After I will again to edit
file structure.

Header and Footer components added on project. I started to create
header's design.